rename direcory of the executor
